Add metadata to parsed file columns

+++ porcino_core/src/network/mod.rs
@@ -57,10 +57,6 @@ impl Network {
                 .map(|(nb, dnb)| nb + dnb)
                 .collect();
 
-            // Something wrong here!!!
-            //dbg!(&nabla_w);
-            //dbg!(&delta_nabla_w);
-            //dbg!(&self);
             nabla_w = nabla_w
                 .iter()
                 .zip(delta_nabla_w.iter())
@@ -74,7 +70,7 @@ impl Network {
             .map(|layer| &mut layer.weights)
             .zip(nabla_w.iter())
             .for_each(|(w, nw)| {
-                *w = w.clone() - nw * (eta as f64);
+                *w = w.clone() - nw * eta;
             });
 
         self.layers
@@ -82,7 +78,7 @@ impl Network {
             .map(|layer| &mut layer.biases)
             .zip(nabla_b.iter())
             .for_each(|(w, nw)| {
-                *w = w.clone() - nw * (eta as f64);
+                *w = w.clone() - nw * eta;
             });
     }
     pub fn calculate_gradient(
@@ -106,10 +102,6 @@ impl Network {
         // Last layer
         let mut delta = (&self.layers.last().unwrap().state - reference_set)
             * &Sigmoid::derivative(&self.layers.last().unwrap().zs, None);
-
-        //dbg!(reference_set - &self.layers.last().unwrap().state);
-        //dbg!(&self.layers.last().unwrap().zs);
-        //dbg!(&delta);
 
         *nabla_b.last_mut().unwrap() = delta.clone();
         *nabla_w.last_mut().unwrap() = if self.layers.len() >= 2 {
